ENH: Add new functions required to use this library for Monte Carlo simulations of proton transport
When performing Monte Carlo simulations of proton therapy treatments,
one needs to track the trajectory of protons in different coordinate systems
of the IEC standard.
For example, you start with the proton in the gantry, directed towards
the isocenter, which is expressed in the DICOM coordinate system. One needs
thus to transform from DICOM into Patient system, and then to Gantry system.
Furthermore, simulation analyzes data from a regular grid derived from the CT,
so one needs to convert also from DICOM coordinate system to the
regular grid, ie to get the voxel indices in 3D according to the imagePosition
DICOM tag as well as the imageOrientation DICOM tag, and also the PixelSpacing
and SliceSeparation attributes.

Thus, this commit adds the required transform hierarchy to handle this mechanisms,
and implements updated function singatures in the header to configure these
matrices according to the parameters defined in the DICOM CT files as well as
DICOM RTplans. This vtkIEC library can then fully support this other downstream
Monte Carlo simulation library that is independent on 3D Slicer.

The resulting transformations have been tested and compared with sympy
python code.

Summary:
1. Updated the Gantry transformation to account for pitch (NOTE: order
was chosen to be pitch rotation (around X) followed by gantry
rotation(around Y). This is a DICOM extension of the IEC standard, compliant with it.
2. Added Patient image plane transformation
3. Added DICOM transformation step from Patient->DICOM->PatientImageRegularGrid
4. Added comprehensive doxygen documentation to all functions.